finished base site

+++ laravel-blog/app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -67,6 +77,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -96,8 +107,13 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        //
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -135,6 +151,10 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+        // check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
